+++ python/metatensor_core/metatensor/_data/__init__.py
@@ -1,17 +1,19 @@
 from ._array import (  # noqa: F401
+    Array,
     Device,
-    DeviceWarning,
     DType,
+    array_backend,
     array_change_backend,
     array_change_device,
     array_change_dtype,
     array_device,
     array_device_is_cpu,
     array_dtype,
+    array_dtype_is_integer,
     create_mts_array,
+    to_arguments_parse,
 )
 from ._extract import (  # noqa: F401
-    Array,
     ExternalCpuArray,
     ExternalCudaArray,
     data_origin,