+++ python/metatensor_core/metatensor/_data/_extract.py
@@ -1,5 +1,5 @@
 import ctypes
-from typing import Any, NewType, Union
+from typing import Any
 
 import numpy as np
 
@@ -21,39 +21,6 @@ from ._array import (
     _register_origin,
 )
 from ._dlpack import DLPackArray, wrap_versioned_as_unversioned
-
-
-try:
-    import torch
-
-    HAS_TORCH = True
-except ImportError:
-    HAS_TORCH = False
-
-
-if HAS_TORCH:
-    # This NewType is only used for typechecking and documentation purposes. If you are
-    # trying to add support for new array types, see `data.array.ArrayWrapper` instead.
-    Array = NewType("Array", Union[np.ndarray, torch.Tensor])
-else:
-    Array = NewType("Array", np.ndarray)
-
-Array.__doc__ = """
-An ``Array`` contains the actual data stored in a :py:class:`metatensor.TensorBlock`.
-
-This data is manipulated by ``metatensor`` in a completely opaque way: this library does
-not know what's inside the arrays appart from a small set of constrains:
-
-- the array contains numeric data (:py:func:`metatensor.load` and
-  :py:func:`metatensor.save` additionally requires contiguous arrays of 64-bit IEEE-754
-  floating points numbers);
-- they are stored as row-major, n-dimensional arrays with at least 2 dimensions;
-- it is possible to create new arrays and move data from one array to another.
-
-The actual type of an ``Array`` depends on how the :py:class:`metatensor.TensorBlock`
-was created. Currently, :py:class:`numpy.ndarray` and :py:class:`torch.Tensor` are
-supported.
-"""
 
 
 _ADDITIONAL_ORIGINS = {}